requirements.txt
Flask
pymupdf4llm
pymupdfpro
fitz